requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
python-multipart==0.0.20
pyyaml==6.0.3
langdetect==1.0.9